add my classes with javadoc - pass test's of level 1

+++ src/geometries/Cylinder.java
@@ -7,7 +7,7 @@ import primitives.Vector;
 public class Cylinder extends Tube{
     double height;
 
-    public Cylinder(float radius, Ray axisRay, double height) {
+    public Cylinder(double radius, Ray axisRay, double height) {
         super(radius, axisRay);
         this.height = height;
     }
