+++ src/geometries/RadialGeometry.java
@@ -3,7 +3,7 @@ package geometries;
 public abstract class RadialGeometry implements Geometry{
     protected double radius;
 
-    public RadialGeometry(float radius) {
+    public RadialGeometry(double radius) {
         this.radius = radius;
     }
 }
