+++ src/geometries/Tube.java
@@ -7,7 +7,7 @@ import primitives.Vector;
 public class Tube extends RadialGeometry{
     Ray axisRay;
 
-    public Tube(float radius, Ray axisRay) {
+    public Tube(double radius, Ray axisRay) {
         super(radius);
         this.axisRay = axisRay;
     }
